multimodel/requirements.txt
antlr4-python3-runtime==4.9.3
hydra-core==1.3.2
numpy==2.1.3
omegaconf==2.3.0
packaging==24.2
pandas==2.2.3
PuLP==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
six==1.16.0
tzdata==2024.2
